requirements.txt
gradio
matplotlib
numpy
scikit-learn
tensorflow
psycopg2-binary

